\name{AdditionTCFIG1}
\alias{AdditionTCFIG1}
\title{ addition tcfig1}
\description{
Class \code{AdditionTCFIG1} from package \code{zorg}.
}
\usage{
AdditionTCFIG1()
}
\value{
An object of class AdditionTCFIG1.
}
\references{
Refer to \href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing R extensions} to know more about R documentation requirements.
}
\author{
\packageAuthor{zorg}

Maintainer: \packageMaintainer{zorg}
}
\examples{
# no tests instrumentation
}
\keyword{classes}
\keyword{environment}
\keyword{utilities}
\keyword{misc}
\concept{evaluation mode}
\concept{standard evaluation}
\concept{function return type evaluation}
\concept{parameter check evaluation}
