\name{zorg-package}
\alias{zorg-package}
\alias{zorg}
\docType{package}
\title{\packageTitle{zorg}}
\description{
\packageDescription{zorg}
}
\details{
Most important package entries are
\itemize{
\item{\code{\link{AdditionTCFIG1}}}{}
\item{\code{\link{AdditionTCFIP}}}{}
\item{\code{\link{Addition}}}{}
}
}
\references{
See \code{wyz.code.offensiveProgramming} package documentation.

You may read \href{https://neonira.github.io/offensiveProgrammingBook/}{Offensive Programming Book} to get introduction and expert advices on offensive programming.

Refer to \href{https://cran.r-project.org/doc/manuals/R-exts.html}{Writing R extensions} to know more about R documentation requirements.
}
\author{
\packageAuthor{zorg}

Maintainer: \packageMaintainer{zorg}
}
\keyword{classes}
\keyword{environment}
\keyword{utilities}
\keyword{misc}
\concept{evaluation mode}
\concept{standard evaluation}
\concept{function return type evaluation}
\concept{parameter check evaluation}
